01-docker-terraform/docker/pyproject.toml
[project]
name = "01-docker-terraform"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "click>=8.3.1",
    "panda>=0.3.1",
    "pyarrow>=23.0.0",
    "sqlalchemy>=2.0.45",
]
